iteration_1/requirements.txt
# Core AI/ML Dependencies for RAG System
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
openai>=1.10.0

# Vector Database (simplified)
chromadb==0.4.15

# Data Processing (compatible versions)
pandas==2.0.3
numpy==1.24.4

# Web Framework for API
fastapi==0.100.0
uvicorn==0.23.0

# Environment Management
python-dotenv==1.0.0

# Utility
pydantic==2.5.0 